badges/requirements.txt
Django==5.0.4
opencv-python-headless==4.9.0.80
pillow==10.3.0
rembg==2.0.56